# docs/swagger.yaml
basePath: /v1/users
definitions:
...
  github_com_Flikest_PingVi_backend_internal_delivery_dto.CreateTopic:
    properties:
      channel_id:
        type: string
      is_private:
        type: boolean
      name:
        type: string
      position:
        type: integer
      topic_type:
        type: string
      user_id:
        type: string
    type: object
...
  github_com_Flikest_PingVi_backend_internal_delivery_dto.DeleteTopic:
    properties:
      channel_id:
        type: string
      id:
        type: string
      user_id:
        type: string
    type: object
...
  github_com_Flikest_PingVi_backend_internal_delivery_dto.Topic:
    properties:
      channel_id:
        type: string
      created_at:
        type: string
      directory_id:
        type: string
      id:
        type: string
      is_private:
        type: boolean
      name:
        type: string
      position:
        type: integer
      topic_type:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_Flikest_PingVi_backend_internal_delivery_dto.UpdateTopic:
    properties:
      channel_id:
        type: string
      id:
        type: string
      is_private:
        type: boolean
      name:
        type: string
      position:
        type: integer
      topic_type:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is PingVi selective forwarding unit.
  termsOfService: http://swagger.io/terms/
  title: PingVi selective forwarding unit
  version: "1.0"
paths:
...
  /topics:
    delete:
      consumes:
      - application/json
      description: |-
        Permanently deletes a topic from a channel. Requires appropriate permissions.
        - System message will be sent to the channel
      parameters:
      - description: Bearer JWT token
        example: '"Bearer eyJhbGciOiJIUzI1NiIs..."'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Topic deletion data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Flikest_PingVi_backend_internal_delivery_dto.DeleteTopic'
      produces:
      - application/json
      responses:
        "200":
          description: "Topic ID\"\t\t\t\texample({\"topic_id\":\"123e4567-e89b-12d3-a456-426614174000\"})"
          schema:
            additionalProperties: true
            type: object
        "400":
          description: "Invalid request\"\t\texample({\"error\":\"invalid body\"})"
          schema:
            additionalProperties: true
            type: object
        "500":
          description: "Internal server error\"\texample({\"error\":\"failed to delete
            topic\"})"
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a topic
      tags:
      - messenger-topics
    post:
      consumes:
      - application/json
      description: |-
        Creates a new topic (text or voice) in a channel. Requires appropriate permissions.
        - Topic types: "text" or "voice"
        - System message will be sent to the channel
      parameters:
      - description: Bearer JWT token
        example: '"Bearer eyJhbGciOiJIUzI1NiIs..."'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Topic creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Flikest_PingVi_backend_internal_delivery_dto.CreateTopic'
      produces:
      - application/json
      responses:
        "201":
          description: Created topic
          schema:
            $ref: '#/definitions/github_com_Flikest_PingVi_backend_internal_delivery_dto.Topic'
        "400":
          description: "Invalid request\"\t\texample({\"error\":\"invalid body\"})"
          schema:
            additionalProperties: true
            type: object
        "500":
          description: "Internal server error\"\texample({\"error\":\"failed to create
            topic\"})"
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new topic in a channel
      tags:
      - messenger-topics
    put:
      consumes:
      - application/json
      description: |-
        Updates an existing topic in a channel. Requires appropriate permissions.
        - Can update: name, topic type
      parameters:
      - description: Bearer JWT token
        example: '"Bearer eyJhbGciOiJIUzI1NiIs..."'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Topic update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Flikest_PingVi_backend_internal_delivery_dto.UpdateTopic'
      produces:
      - application/json
      responses:
        "200":
          description: Updated topic
          schema:
            $ref: '#/definitions/github_com_Flikest_PingVi_backend_internal_delivery_dto.Topic'
        "400":
          description: "Invalid request\"\t\texample({\"error\":\"invalid body\"})"
          schema:
            additionalProperties: true
            type: object
        "500":
          description: "Internal server error\"\texample({\"error\":\"failed to update
            topic\"})"
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a topic
      tags:
      - messenger-topics
  /topics/{channel_id}:
    get:
      consumes:
      - application/json
      description: Retrieves all topics from a specific channel. User must be a member
        of the channel.
      parameters:
      - description: Bearer JWT token
        example: '"Bearer eyJhbGciOiJIUzI1NiIs..."'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Channel ID
        example: '"123e4567-e89b-12d3-a456-426614174000"'
        in: path
        name: channel_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of topics
          schema:
            items:
              $ref: '#/definitions/github_com_Flikest_PingVi_backend_internal_delivery_dto.Topic'
            type: array
        "400":
          description: "Invalid channel ID\"\texample({\"error\":\"invalid channel
            id\"})"
          schema:
            additionalProperties: true
            type: object
        "403":
          description: "Not a member\"\t\t\texample({\"error\":\"the user does not
            belong to this group\"})"
          schema:
            additionalProperties: true
            type: object
        "500":
          description: "Internal server error\"\texample({\"error\":\"failed to get
            topics\"})"
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all topics in a channel
      tags:
      - messenger-topics
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"